requirements.txt
pandas 
sqlalchemy 
psycopg2
python-dotenv